+++ src/routes/Home.js
@@ -2,12 +2,12 @@ import { useEffect } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { fetchUser } from "../redux/actions/currentUserAction";
 import Navbar from "../components/Navbar"
+import Header from "../components/Header"
 import "react-toastify/dist/ReactToastify.css";
 
 const Home = () => {
   const dispatch = useDispatch();
 
-  const currentUser = useSelector((state) => state.currentUser.userData);
   const status = useSelector((state) => state.currentUser.status);
   const error = useSelector((state) => state.currentUser.error);
 
@@ -20,21 +20,14 @@ const Home = () => {
   }
 
   if (status === "failed") {
-    return <div>Error: {error}</div>;
+    return <div>Please Connect again..{error}</div>;
   }
 
   return (
     <div className="flex w-full h-screen">
       <Navbar />
       <div className="bg-slate-900 flex flex-col w-[80%] text-teal-400 text-white">
-        <header className="flex justify-end py-4 pr-4">
-          {currentUser && (
-            <button type="button" className="flex gap-4 bg-teal-400 p-2 rounded">
-              <p className="font-bold text-slate-900">{currentUser.userName}</p>
-              <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24"><path fill="text-slate-900" d="M12 4a4 4 0 0 1 4 4a4 4 0 0 1-4 4a4 4 0 0 1-4-4a4 4 0 0 1 4-4m0 10c4.42 0 8 1.79 8 4v2H4v-2c0-2.21 3.58-4 8-4"/></svg>
-            </button>
-          )}
-        </header>
+        <Header />
         <div className="h-screen">home</div>
       </div>
     </div>
